Better docs (#7)
**Story card:** [sc-7751](https://app.shortcut.com/simpledotorg/story/7751/add-any-missing-documentation)

## This addresses

Documents all configuration variables in both `vars.yml` and `vault.yml`, with the idea that if a third party was running Simple, they'd be able to understand what each of the configs do, and how to change them.

Note: I've only edited ET Demo for now. After review approval, I'll port the same documentation to ET prod too. I plan to create a generator and some templates in [an upcoming story](https://app.shortcut.com/simpledotorg/story/7970/create-generator-for-new-deployment), and those templates should serve as the source of truth for this documentation.