fix: Set the environment variables the Ansible way (#73)
The original change assumed that commands would be as is done in shell
scripts. In Ansible, the command is a string. Environment variables
tacked on to the command would be read as part of the command.

This change ensures that we set environment variables the way Ansible
expects they be set. We still want to keep `rails` as the task runner
here, since `rake` is has been phased out.